requirements.txt
alembic==1.11.2
fastapi==0.101.0
psycopg2==2.9.7
uvicorn[standard]==0.23.2
sqlmodel==0.0.8
pytest==7.4.0
requests==2.31.0
pytest-cov==4.1.0
pytest-factoryboy==2.5.1
black==23.7.0
flake8==6.1.0
isort==5.12.0
pylint==2.17.5
mypy==1.4.1
types-requests==2.31.0.2
sqlalchemy-utils==0.41.1
gunicorn==21.2.0
typer[all]==0.9.0
python-ldap==3.4.3
python-multipart==0.0.6
python-dateutil==2.8.2
types-python-dateutil==2.8.19.14
freezegun==1.2.2
fastapi-cache2[redis]==0.2.1
aioredis==2.0.1
click==8.1.6
httpx==0.24.1

